writing output is now handeled in new class
